Assignment_4/requirements.txt
gym==0.21.0
imageio==2.14.0
ipython==8.0.1
matplotlib==3.5.1
numpy==1.21.4
Pillow==9.0.0
scikit_learn==1.0.2
tensorflow==2.7.0
tf_agents==0.11.0
